extras: move peripherals from nrf to extras
