+++ embassy-extras/src/lib.rs
@@ -3,6 +3,7 @@
 // This mod MUST go first, so that the others see its macros.
 pub(crate) mod fmt;
 
+mod macros;
 pub mod peripheral;
 pub mod ring_buffer;
