+++ embassy-nrf/src/peripherals.rs
@@ -1,51 +1,4 @@
-use embassy::util::PeripheralBorrow;
-
-macro_rules! peripherals {
-    ($($(#[$cfg:meta])? $name:ident: $type:ident),*$(,)?) => {
-        $(
-            $(#[$cfg])?
-            #[allow(non_camel_case_types)]
-            pub struct $type { _private: () }
-
-            $(#[$cfg])?
-            impl PeripheralBorrow for $type {
-                type Target = $type;
-                unsafe fn unborrow(self) -> $type {
-                    self
-                }
-            }
-
-            $(#[$cfg])?
-            impl PeripheralBorrow for &mut $type {
-                type Target = $type;
-                unsafe fn unborrow(self) -> $type {
-                    ::core::ptr::read(self)
-                }
-            }
-        )*
-
-        pub struct Peripherals {
-            $(
-                $(#[$cfg])?
-                pub $name: $type,
-            )*
-        }
-
-        impl Peripherals {
-            pub unsafe fn steal() -> Self {
-                Self {
-                    $(
-                        $(#[$cfg])?
-                        $name: $type { _private: () },
-                    )*
-                }
-            }
-        }
-
-    };
-}
-
-peripherals! {
+embassy_extras::peripherals! {
     // RTC
     rtc0: RTC0,
     rtc1: RTC1,
